+++ ml/jml/classifier.h
@@ -426,12 +426,12 @@ public:
         given way.  If not implemented, an exception will be thrown.
     */
     virtual Explanation explain(const Feature_Set & feature_set,
-                                int label,
+                                float label,
                                 double weight = 1.0,
                                 PredictionContext * context = 0) const;
 
     virtual Explanation explainContext(const Feature_Set & feature_set,
-                                       int label,
+                                       float label,
                                        const PredictionContext & context,
                                        double weight = 1.0) const
     {
