+++ ml/jml/decision_tree.h
@@ -114,13 +114,13 @@ public:
                                 double weight = 1.0) const;
 
     virtual Explanation explain(const Feature_Set & feature_set,
-                                int label,
+                                float label,
                                 double weight = 1.0,
                                 PredictionContext * context = 0) const;
 
     void explain_recursive(Explanation & explanation,
                            const Feature_Set & fset,
-                           int label,
+                           float label,
                            double weight,
                            const Tree::Ptr & ptr,
                            const Tree::Node * parent) const;
